+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.4.2] - 2019-01-05
+### Added
+- Added `Shortcode::with_injector()` method to allow for the use of an external injector.
+
+### Changed
+- Changed the way `$context` is handled to make it easier to modify from external code, especially a DI.
+
 ## [0.4.1] - 2019-01-05
 ### Added
 - Added `Shortcode::add_context()` method.
@@ -152,6 +159,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release to GitHub.
 
+[0.4.2]: https://github.com/brightnucleus/shortcodes/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/brightnucleus/shortcodes/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/brightnucleus/shortcodes/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/brightnucleus/shortcodes/compare/v0.3.1...v0.3.2